Move shared helper functions to separate util module
