+++ tests/agents/test_fakedata.py
@@ -2,35 +2,13 @@ import sys
 sys.path.insert(0, '../agents/fake_data/')
 from fake_data_agent import FakeDataAgent
 
-import pytest
 import pytest_twisted
-from unittest import mock
 
-from ocs.ocs_agent import OpSession
-
-import txaio
-txaio.use_twisted()
+from util import create_session, create_agent_fixture
 
 
-@pytest.fixture
-def agent():
-    """Test fixture to setup a mocked OCSAgent."""
-    mock_agent = mock.MagicMock()
-    log = txaio.make_logger()
-    txaio.start_logging(level='debug')
-    mock_agent.log = log
-    log.info('Initialized mock OCSAgent')
-    agent = FakeDataAgent(mock_agent)
-
-    return agent
-
-
-def create_session(op_name):
-    """Create an OpSession with a mocked app for testing."""
-    mock_app = mock.MagicMock()
-    session = OpSession(1, op_name, app=mock_app)
-
-    return session
+# fixtures
+agent = create_agent_fixture(FakeDataAgent)
 
 
 def test_fake_data_set_heartbeat(agent):
@@ -73,8 +51,6 @@ def test_fake_data_delay_task(agent):
 
 
 def test_fake_data_try_set_job_running_job(agent):
-    session = create_session('acq')
-
     # set running job to 'acq'
     agent.job = 'acq'
